+++ examples/izhikevich.py
@@ -33,7 +33,7 @@ u = u + d
 """
 '''
 
-model = build(DSL, stepper_name="euler", jit=False, model_dtype="float32")
+model = build(DSL, stepper="euler", jit=False, dtype="float32")
 sim = Sim(model)
 sim.run(t_end=600.0, dt=0.01, cap_rec=10000)